requirements.txt
affine==2.4.0
aio-pika==9.3.0
aioamqp==0.15.0
aiormq==6.7.7
alteia==2.2.0
amqp==5.1.1
anyio==4.0.0
appdirs==1.4.4
APScheduler==3.9.1
area==1.1.1
asgiref==3.5.0
async-timeout==4.0.3
attrs==23.1.0
autopep8==1.6.0
Babel==2.12.1
bcrypt==3.2.0
beautifulsoup4==4.12.2
billiard==4.1.0
black==22.3.0
blinker==1.6.3
branca==0.5.0
Brotli==1.0.9
cachetools==4.2.4
carehare==1.0.2
celery==5.3.4
certifi==2021.10.8
cffi==1.15.0
channels==4.0.0
channels-rabbitmq==4.0.1
channels-redis==4.1.0
charset-normalizer==2.0.12
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cligj==0.7.2
colorama==0.4.4
ConfigArgParse==1.7
confluent-kafka==2.2.0
cryptography==36.0.1
cssselect2==0.5.0
defusedxml==0.7.1
dill==0.3.7
dj-database-url==0.5.0
Django==4.0.5
django-allauth==0.51.0
django-heroku==0.3.1
django-phonenumber-field==7.1.0
django-webpack-loader==3.0.1
earthengine-api==0.1.312
et-xmlfile==1.1.0
exceptiongroup==1.1.3
Fiona==1.9.4.post1
Flask==3.0.0
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
folium==0.12.1.post1
fonttools==4.31.2
future==0.18.2
fuzzywuzzy==0.18.0
geojson==2.5.0
geopandas==0.13.2
gevent==23.9.1
geventhttpclient==2.0.11
git-filter-repo==2.38.0
google-api-core==1.34.0
google-api-python-client==1.12.1
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.2
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.3.0
googleapis-common-protos==1.60.0
greenlet==3.0.1
gunicorn==20.1.0
h11==0.14.0
html5lib==1.1
httplib2==0.20.4
httplib2shim==0.0.3
httptools==0.6.1
idna==3.3
importlib-metadata==6.7.0
itsdangerous==2.1.2
jellyfish==0.9.0
Jinja2==3.1.2
joblib==1.3.2
kombu==5.3.2
Levenshtein==0.20.9
locust==2.18.0
MarkupSafe==2.1.3
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==0.4.3
mysqlclient==2.1.0
numpy==1.22.4
oauth2client==4.1.3
oauthlib==3.2.0
openpyxl==3.0.9
ormsgpack==1.4.1
packaging==23.1
pamqp==3.2.1
pandas==1.4.2
paramiko==2.11.0
pathspec==0.9.0
pathvalidate==2.5.0
phonenumbers==8.13.14
pika==1.3.2
Pillow==9.0.1
platformdirs==2.5.2
prompt-toolkit==3.0.28
protobuf==3.20.3
psutil==5.9.6
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydyf==0.1.2
PyJWT==2.4.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==3.0.7
pyphen==0.12.0
pyproj==3.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-Levenshtein==0.20.9
python3-openid==3.2.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
pyzmq==25.1.1
rapidfuzz==3.5.2
rasterio==1.3.8
redis==5.0.1
reportlab==3.6.10
requests==2.27.1
requests-oauthlib==1.3.1
roundrobin==0.0.4
rsa==4.8
semantic-version==2.9.0
Shapely==1.8.2
six==1.16.0
sniffio==1.3.0
snuggs==1.4.7
soupsieve==2.5
SQLAlchemy==2.0.22
sqlparse==0.4.2
sshtunnel==0.4.0
thefuzz==0.20.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
tzdata==2023.3
tzlocal==4.1
Unidecode==1.3.4
uritemplate==3.0.1
urllib3==1.26.8
uvicorn==0.23.2
uvloop==0.19.0
vine==5.0.0
watchfiles==0.21.0
wcwidth==0.2.5
weasyprint==54.2
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.1
whitenoise==6.1.0
xlwt==1.3.0
yarl==1.9.2
zipp==3.15.0
zope.event==5.0
zope.interface==6.1
zopfli==0.2.1
